requirements.txt
streamlit
requests
beautifulsoup4
sentence-transformers
transformers
torch
faiss-cpu
numpy

